third_party/WebKit/LayoutTests/fast/notifications/notification-properties-expected.txt

Checks that the {title, dir, lang, body, tag, icon} properties are set on the Notification object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS notification.title is "My Notification"
PASS notification.dir is "rtl"
PASS notification.lang is "nl-NL"
PASS notification.body is "Hallo, wereld!"
PASS notification.tag is "notification"
PASS notification.icon is "http://localhost/my_icon.png"
PASS successfullyParsed is true

TEST COMPLETE
